test: move ws upgrade-event tests to a dedicated file
The upgrade-event tests were in ws.test.ts alongside tests that spawn a
subprocess echo server per case. Those subprocess tests time out under the
slow ASAN debug build (1000ms timeout vs >2s cold spawn), so the file never
reaches a clean pass there regardless of the fix. Move the upgrade-event
tests into ws-upgrade.test.ts, which uses only in-process servers and runs
fast under ASAN.
